Split path_server into a module per class
