+++ bin/path_server
@@ -1,0 +1,21 @@
+#!/usr/bin/python3
+# Copyright 2014 ETH Zurich
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# SCION
+from infrastructure.path_server.core import CorePathServer
+from infrastructure.path_server.local import LocalPathServer
+from lib.main import main_default, main_wrapper
+
+main_wrapper(main_default, CorePathServer, LocalPathServer)
