--- sphinx-doc/infrastructure/path_server.rst
@@ -1,5 +1,0 @@
-
-.. automodule:: infrastructure.path_server
-   :special-members: __init__
-   :members:
-
