+++ sphinx-doc/infrastructure/path_server/base.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.path_server.base
+   :special-members: __init__
+   :members:
+
