+++ sphinx-doc/infrastructure/path_server/core.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.path_server.core
+   :special-members: __init__
+   :members:
+
