+++ sphinx-doc/infrastructure/path_server/index.rst
@@ -1,0 +1,10 @@
+Path Server
+===========
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
