+++ sphinx-doc/infrastructure/path_server/local.rst
@@ -1,0 +1,5 @@
+
+.. automodule:: infrastructure.path_server.local
+   :special-members: __init__
+   :members:
+
